requirements.txt
appdirs==1.4.2
defusedxml==0.5.0
Django==1.10.5
django-allauth==0.30.0
django-file-form==0.2.0
django-form-utils==1.0.3
easy-thumbnails==2.3
oauthlib==2.0.1
olefile==0.44
packaging==16.8
pathlib==1.0.1
Pillow==4.0.0
pyparsing==2.1.10
python3-openid==3.1.0
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
